refactor: 패키지명 및 그룹명 변경
