+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'fintory-backend-project'
+rootProject.name = 'fintory-BE'
